+++ app/src/main/java/com/stonks/ui/Constants.kt
@@ -1,0 +1,7 @@
+package com.stonks.ui
+
+class Constants {
+    companion object {
+        const val defaultCurrencyInd = 17 // Index of USD
+    }
+}
