+++ app/src/main/res/layout/activity_onboarding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnboardingActivity">
+    tools:context=".ui.OnboardingActivity">
 
     <TextView
         android:layout_width="wrap_content"
